feat: finisht  to verify if the person who already login is the same person that try to update or delete

+++ src/pessoas/pessoas.service.ts
@@ -1,5 +1,6 @@
 import {
   ConflictException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
   Scope,
@@ -84,6 +85,9 @@ export class PessoasService {
     if (!pessoa) {
       throw new NotFoundException('Pessoa não encontrada');
     }
+    if (pessoa.id !== tokenPayload.sub) {
+      throw new ForbiddenException('Você não é essa pessoa.');
+    }
     return this.pessoaRepository.save(pessoa);
   }
 
@@ -91,6 +95,9 @@ export class PessoasService {
     const pessoa = await this.findOne(id);
     if (!pessoa) {
       throw new NotFoundException('Pessoa não encontrada');
+    }
+    if (pessoa.id !== tokenPayload.sub) {
+      throw new ForbiddenException('Você não é essa pessoa.');
     }
     return await this.pessoaRepository.remove(pessoa);
   }
